+++ akshare/__init__.py
@@ -3202,9 +3202,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.17.93 fix: fix stock_news_em interface
 1.17.94 fix: fix index_detail_cni interface
 1.17.95 fix: fix stock_info_sh_name_code interface
+1.17.96 fix: fix stock_info_sh_name_code interface
 """
 
-__version__ = "1.17.95"
+__version__ = "1.17.96"
 __author__ = "AKFamily"
 
 import sys