+++ docs/changelog.md
@@ -89,6 +89,14 @@
 
 ## 更新说明详情
 
+1.17.96 fix: fix stock_info_sh_name_code interface
+
+	1. 修复 stock_info_sh_name_code 接口
+	2. 修复 stock_financial_cash_ths 接口
+	3. 修复 stock_financial_benefit_ths 接口
+	4. 修复 stock_financial_debt_ths 接口
+	5. 修复 stock_financial_abstract_ths 接口
+
 1.17.95 fix: fix stock_info_sh_name_code interface
 
 	1. 修复 stock_info_sh_name_code 接口
@@ -5996,6 +6004,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.17.96 fix: fix stock_info_sh_name_code interface
 
 1.17.95 fix: fix stock_info_sh_name_code interface